+++ app/src/main/res/layout/activity_edit_notes.xml
